New analyses scripts and data exploration
Finalized models for logistic mixed models and AIC model selection to
examine effect of PAs and habitat use on presence/absence. Added a script
specific for running this on the HPC Wales cluster. Separated scripts
plotting out the data for analyses for exploration vs the script running
the actual models. Started to add models looking at the effect of
PA designation level on use. Created a source code to take a list of GLMM
model outputs and do AIC model selection on it - outputs AIC table with
model weights.